SV: Implement RBAC, staff page

+++ cctv_manager/cameras/views.py
@@ -1,4 +1,6 @@
-from django.http import HttpResponseRedirect, JsonResponse
+from typing import Any, Dict
+
+from django.http import HttpResponseRedirect, JsonResponse, HttpResponse
 from django.urls import reverse_lazy
 from django.views.generic import DetailView, CreateView, UpdateView
 from django_tables2 import SingleTableView
@@ -6,19 +8,47 @@ from django.shortcuts import get_object_or_404
 
 from . import forms, models, tables
 from utils.network_scripts import check_ping
+from utils.rbac_scripts import is_user_able
 
 
-class CamerasListView(SingleTableView):
+class CameraViewFilterMixin:
+    def get_queryset(self):
+        return models.Camera.objects.for_user(self.request.user)
+
+
+class CameraChangeFilterMixin:
+    def get_queryset(self):
+        return models.Camera.objects.for_user(self.request.user, action='C')
+
+
+class CameraAddFilterMixin:
+    def get_queryset(self):
+        return models.Camera.objects.for_user(self.request.user, action='A')
+
+
+class CameraDeleteFilterMixin:
+    def get_queryset(self):
+        return models.Camera.objects.for_user(self.request.user, action='D')
+
+
+class CamerasListView(CameraViewFilterMixin, SingleTableView):
     model = models.Camera
     template_name = 'cameras/list.html'
     table_class = tables.CamerasTable
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update({
+            'is_able_add': is_user_able(self.request.user, 'C', 'A'),
+        })
+        return context
 
     def get_queryset(self):
         fields = ['name', 'ip_address']
         return super().get_queryset().only(*fields)
 
 
-class CameraDetailView(DetailView):
+class CameraDetailView(CameraViewFilterMixin, DetailView):
     model = models.Camera
     template_name = 'cameras/detail.html'
 
@@ -27,12 +57,14 @@ class CameraDetailView(DetailView):
         camera = self.object
         context.update({
             'fields': camera.get_fields().items(),
-            # 'status': check_ping(camera.ip_address),
+            'is_able_edit': is_user_able(self.request.user, 'C', 'C'),
+            'is_able_delete': is_user_able(self.request.user, 'C', 'D'),
+            'is_able_reload': is_user_able(self.request.user, 'C', 'R'),
         })
         return context
 
 
-class CameraAddView(CreateView):
+class CameraAddView(CameraAddFilterMixin, CreateView):
     model = models.Camera
     template_name = 'cameras/create.html'
     form_class = forms.CameraAddForm
@@ -41,7 +73,7 @@ class CameraAddView(CreateView):
         return reverse_lazy('camera_detail', kwargs={'pk': self.object.id})
 
 
-class CameraEditView(UpdateView):
+class CameraEditView(CameraChangeFilterMixin, UpdateView):
     model = models.Camera
     template_name = 'cameras/edit.html'
     fields = ['name', 'ip_address', 'prerecord_time_sec', 'postrecord_time_sec', 'video_length_min']
@@ -51,6 +83,8 @@ class CameraEditView(UpdateView):
 
 
 def camera_delete(request, pk):
+    if not is_user_able(request.user, 'C', 'D'):
+        return HttpResponse(status=404)
     camera = get_object_or_404(models.Camera, pk=pk)
     camera.delete()
     return HttpResponseRedirect(reverse_lazy('cameras_list'))
